% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2ListBuckets.R
\name{b2ListBuckets}
\alias{b2ListBuckets}
\title{B2 List Buckets.}
\usage{
b2ListBuckets()
}
\value{
If successful a data frame will be returned containing
  \code{accountId}, \code{bucketId}, \code{bucketName} and \code{bucketType}
  for all existing buckets in the specific user's account.
}
\description{
\code{b2ListBuckets} lists buckets associated with an account, in
alphabetical order by bucket ID.
}
\details{
This function lists buckets associated with a user's account on the Backblaze
B2 cloud storage product, in alphabetical order by bucket ID. Further details
regarding this API call are available here:

\url{https://www.backblaze.com/b2/docs/b2_list_buckets.html}

B2 List Buckets does not support any function options. However, a the
function must be authorised through the \code{b2AuthorizeAccount} function
first.
}
\examples{
\dontrun{
b2ListBuckets()
}

}

